flexi_logger

Function init

pub fn init()

Shortest form to get started.

flexi_logger::init();.

Equivalent to

    Logger::try_with_env_or_str("info")
       .unwrap_or_else(|_e| Logger::with(LogSpecification::info()))
       .log_to_stderr()
       .start()
       .ok();

that means,

you configure the log specification via the environment variable RUST_LOG, or use the default log specification ('info')

logs are directly written to stderr, without any buffering, so implicitly dropping the LogHandle (which is returned from Logger::start()) is ok.