Function flexi_logger::style

pub fn style<T>(level: Level, item: T) -> Paint<T>

Helper function that is used in the provided coloring format functions to apply colors based on the log level and the effective color palette.

See Logger::set_palette if you want to modify the color palette.

Only available with feature colors.