Struct flexi_logger::FlexiLogger

pub struct FlexiLogger { /* fields omitted */ }

Does the logging in the background, is normally not used directly.

This struct is only required if you want to allow supporting multiple FlexiLogger instances in a single process.

Methods

impl FlexiLogger

fn new(
loglevelspec: Option<String>,
config: LogConfig
) -> Result<FlexiLogger, FlexiLoggerError>

Creates a new FlexiLogger instance based on your configuration and a loglevel specification.

fn fl_enabled(&self, level: LogLevel, target: &str) -> bool

Checks if a log line for the specified target and level is to be written really