Struct flexi_logger::LogConfig

pub struct LogConfig {
    pub log_to_file: bool,
    pub print_message: bool,
    pub duplicate_error: bool,
    pub duplicate_info: bool,
    pub format: fn(&LogRecord) -> String,
}

Allows influencing the behavior of the FlexiLogger.

Fields

log_to_file

If true, the log is written to a file. Default is false, the log is then written to stderr. If true, a new file in the current directory is created and written to. The name of the file is chosen as '<program_name>_<date>_<time>.trc', e.g. myprog_2015-07-08_10-44-11.trc

print_message

If true (which is default), and if log_to_file is true, the name of the tracefile is documented in a message to stdout.

duplicate_error

If true (which is default), and if log_to_file is true, all error messages are written also to stdout.

duplicate_info

If true (which is default), and if log_to_file is true, also info messages are written also to stdout.

format

Allows providing a custom logline format; default is flexi_logger::default_format.

Methods

impl LogConfig

fn new() -> LogConfig