Struct flexi_logger::LogSpecBuilder

pub struct LogSpecBuilder { /* private fields */ }

Builder for LogSpecification.

Example

Use the reconfigurability feature and build the log spec programmatically.

use flexi_logger::{Logger, LogSpecBuilder};
use log::LevelFilter;

fn main() {
    // Build the initial log specification
    let mut builder = LogSpecBuilder::new();  // default is LevelFilter::Off
    builder.default(LevelFilter::Info);
    builder.module("karl", LevelFilter::Debug);

    // Initialize Logger, keep builder alive
    let mut logger_reconf_handle = Logger::with(builder.build())
        // your logger configuration goes here, as usual
        .start()
        .unwrap_or_else(|e| panic!("Logger initialization failed with {}", e));

    // ...

    // Modify builder and update the logger
    builder.default(LevelFilter::Error);
    builder.remove("karl");
    builder.module("emma", LevelFilter::Trace);

    logger_reconf_handle.set_new_spec(builder.build());

    // ...
}

Implementations

impl LogSpecBuilder

pub fn new() -> Self

Creates a LogSpecBuilder with all logging turned off.

pub fn from_module_filters(module_filters: &[ModuleFilter]) -> Self

Creates a LogSpecBuilder from given module filters.

pub fn default(&mut self, lf: LevelFilter) -> &mut Self

Adds a default log level filter, or updates the default log level filter.

pub fn module<M: AsRef<str>>(
&mut self,
module_name: M,
lf: LevelFilter
) -> &mut Self

Adds a log level filter, or updates the log level filter, for a module.

pub fn remove<M: AsRef<str>>(&mut self, module_name: M) -> &mut Self

Adds a log level filter, or updates the log level filter, for a module.

pub fn insert_modules_from(&mut self, other: LogSpecification) -> &mut Self

Adds log level filters from a LogSpecification.

pub fn finalize(self) -> LogSpecification

Creates a log specification without text filter.

pub fn finalize_with_textfilter(self, tf: Regex) -> LogSpecification

Creates a log specification with text filter.

This method is only avaible with feature textfilter, which is a default feature.

pub fn build(&self) -> LogSpecification

Creates a log specification without being consumed.

pub fn build_with_textfilter(&self, tf: Option<Regex>) -> LogSpecification

Creates a log specification without being consumed, optionally with a text filter.

This method is only avaible with feature textfilter, which is a default feature.