flexi_logger::writers

Struct FileLogWriterHandle

pub struct FileLogWriterHandle(/* private fields */);

Handle to the FileLogWriter in an ArcFileLogWriter that shuts down the FileLogWriter in its Drop implementation.

Implementations

impl FileLogWriterHandle

pub fn existing_log_files( &self, selector: &LogfileSelector, ) -> Result<Vec<PathBuf>, FlexiLoggerError>

Returns the list of existing log files according to the current FileSpec.

Depending on the given selector, the list may include the CURRENT log file and the compressed files, if they exist.

Errors

FlexiLoggerError::Poison if some mutex is poisoned.