Struct flexi_logger::Logger

pub struct Logger { /* fields omitted */ }

The entry-point for using flexi_logger.

A simple example with file logging might look like this:

use flexi_logger::{Duplicate,Logger};

Logger::with_str("info, mycrate = debug")
        .log_to_file()
        .duplicate_to_stderr(Duplicate::Warn)
        .start()
        .unwrap_or_else(|e| panic!("Logger initialization failed with {}", e));

Logger is a builder class that allows you to

specify your desired (initial) loglevel-specification

either programmatically as a String (Logger::with_str())

or by providing a String in the environment (Logger::with_env()),

or by combining both options (Logger::with_env_or_str()),

or by building a LogSpecification programmatically (Logger::with()),

use the desired configuration methods,

and finally start the logger with

start(),

or start_with_specfile().

Methods

impl Logger

Choose a way to create a Logger instance and define how to access the (initial) loglevel-specification.

pub fn with(logspec: LogSpecification) -> Logger

Creates a Logger that you provide with an explicit LogSpecification. By default, logs are written with default_format to stderr.

pub fn with_str<S: AsRef<str>>(s: S) -> Logger

Creates a Logger that reads the LogSpecification from a String or &str. See LogSpecification for the syntax.

pub fn with_env() -> Logger

Creates a Logger that reads the LogSpecification from the environment variable RUST_LOG.

pub fn with_env_or_str<S: AsRef<str>>(s: S) -> Logger

Creates a Logger that reads the LogSpecification from the environment variable RUST_LOG, or derives it from the given String, if RUST_LOG is not set.

impl Logger

Choose a way how to start logging.

pub fn start(self) -> Result<ReconfigurationHandle, FlexiLoggerError>

Consumes the Logger object and initializes flexi_logger in a way that subsequently the log specification can be updated programmatically.

This allows e.g. to intensify logging for (buggy) parts of a (test) program, etc.

See ReconfigurationHandle for an example.

pub fn start_reconfigurable(
self
) -> Result<ReconfigurationHandle, FlexiLoggerError>

Deprecated since 0.10.6:
please use start() instead

This method is deprecated, because there is no performance advantage any more for the old start() implementation (which did not all roconfiguration).

impl Logger

Simple methods for influencing the behavior of the Logger.

pub fn check_parser_error(self) -> Result<Logger, FlexiLoggerError>

Allows verifying that no parsing errors have occured in the used factory method, and examining the parse error.

The factory methods Logger::with_str(), Logger::with_env(), and Logger::with_env_or_str(), parse a log specification String, and deduce from it a LogSpecification object. Parsing errors are reported to stdout, but effectively ignored; in worst case, a LogSpecification might be used that turns off logging completely!

This method gives programmatic access to parse errors, if there were any.

In the following example we just panic if the spec was not free of errors:

Logger::with_str(some_log_spec_string)
.check_parser_error()
.unwrap()
.log_to_file()
.start();

pub fn log_to_file(self) -> Logger

Makes the logger write all logs to a file, rather than to stderr.

The default pattern for the filename is '<program_name>_<date>_<time>.<suffix>', e.g. myprog_2015-07-08_10-44-11.log.

pub fn do_not_log(self) -> Logger

Makes the logger write no logs at all.

This can be useful when you want to run tests of your programs with all log-levels active to ensure they are working (note that the log macros prevent arguments to inactive log-levels from being evaluated).

pub fn print_message(self) -> Logger

Makes the logger print an info message to stdout with the name of the logfile when a logfile is opened for writing.

pub fn duplicate_to_stderr(self, dup: Duplicate) -> Logger

Makes the logger write messages with the specified minimum severity additionally to stderr.

pub fn duplicate_error(self) -> Logger

Deprecated:
use duplicate_to_stderr(dup: Duplicate)

Makes the logger write all logged error messages additionally to stderr.

pub fn duplicate_info(self) -> Logger

Deprecated:
use duplicate_to_stderr(dup: Duplicate)

Makes the logger write all logged error, warning, and info messages additionally to stderr.

pub fn format(self, format: FormatFunction) -> Logger

Makes the logger use the provided format function for the log entries, rather than formats::default_format.

You can either choose between some predefined variants, default_format, opt_format, detailed_format, with_thread, or you create and use your own format function with the signature fn(&Record) -> String.

pub fn directory<S: Into<String>>(self, directory: S) -> Logger

Specifies a folder for the log files.

This parameter only has an effect if log_to_file() is used, too. If the specified folder does not exist, the initialization will fail. By default, the log files are created in the folder where the program was started.

pub fn suffix<S: Into<String>>(self, suffix: S) -> Logger

Specifies a suffix for the log files.

This parameter only has an effect if log_to_file() is used, too.

pub fn suppress_timestamp(self) -> Logger

Makes the logger not include a timestamp into the names of the log files.

This option only has an effect if log_to_file() is used, too.

pub fn rotate_over_size(self, rotate_over_size: usize) -> Logger

Prevents indefinite growth of log files.

For details, see FileLogWriterBuilder::rotate_over_size.

pub fn append(self) -> Logger

Makes the logger append to the specified output file, if it exists already; by default, the file would be truncated.

This option only has an effect if log_to_file() is used, too. This option will hardly make an effect if suppress_timestamp() is not used.

pub fn discriminant<S: Into<String>>(self, discriminant: S) -> Logger

The specified String is added to the log file name after the program name.

This option only has an effect if log_to_file() is used, too.

pub fn create_symlink<S: Into<String>>(self, symlink: S) -> Logger

The specified String will be used on linux systems to create in the current folder a symbolic link to the current log file.

This option only has an effect if log_to_file() is used, too.

pub fn add_writer<S: Into<String>>(
self,
name: S,
writer: Box<dyn LogWriter>
) -> Logger

Registers a LogWriter implementation under the given target name.

The target name should not start with an underscore.

See the module documentation of writers.

impl Logger

Alternative set of methods to control the behavior of the Logger. Use these methods when you want to control the settings flexibly, e.g. with commandline arguments via docopts or clap.

pub fn o_log_to_file(self, log_to_file: bool) -> Logger

With true, makes the logger write all logs to a file, otherwise to stderr.

pub fn o_print_message(self, print_message: bool) -> Logger

With true, makes the logger print an info message to stdout, each time when a new file is used for log-output.

pub fn o_duplicate_error(self, duplicate_error: bool) -> Logger

Deprecated:
use duplicate_to_stderr(dup: Duplicate)

With true, makes the logger write all logged error messages additionally to stderr; with false, no messages are duplicated.

pub fn o_duplicate_info(self, duplicate_info: bool) -> Logger

Deprecated:
use duplicate_to_stderr(dup: Duplicate)

With true, makes the logger write all logged error, warning, and info messages additionally to stderr; with false, no messages are duplicated.

pub fn o_directory<S: Into<String>>(self, directory: Option<S>) -> Logger

Specifies a folder for the log files.

This parameter only has an effect if log_to_file is set to true. If the specified folder does not exist, the initialization will fail. With None, the log files are created in the folder where the program was started.

pub fn o_rotate_over_size(self, rotate_over_size: Option<usize>) -> Logger

This option only has an effect if log_to_file is set to true.

By default, and with None, the log file will grow indefinitely. If a size is set, when the log file reaches or exceeds the specified size, the file will be closed and a new file will be opened. Also the filename pattern changes: instead of the timestamp, a serial number is included into the filename.

The size is given in bytes, e.g. o_rotate_over_size(Some(1_000)) will rotate files once they reach a size of 1 kB.

pub fn o_timestamp(self, timestamp: bool) -> Logger

With true, makes the logger include a timestamp into the names of the log files. true is the default, but rotate_over_size sets it to false. With this method you can set it to true again.

This parameter only has an effect if log_to_file is set to true.

pub fn o_append(self, append: bool) -> Logger

This option only has an effect if log_to_file is set to true.

If append is set to true, makes the logger append to the specified output file, if it exists. By default, or with false, the file would be truncated.

This option will hardly make an effect if suppress_timestamp() is not used.

pub fn o_discriminant<S: Into<String>>(self, discriminant: Option<S>) -> Logger

This option only has an effect if log_to_file is set to true.

The specified String is added to the log file name.

pub fn o_create_symlink<S: Into<String>>(self, symlink: Option<S>) -> Logger

This option only has an effect if log_to_file is set to true.

If a String is specified, it will be used on linux systems to create in the current folder a symbolic link with this name to the current log file.