Trait flexi_logger::writers::LogWriter

pub trait LogWriter: Sync + Send {
    // Required methods
    fn write(&self, now: &mut DeferredNow, record: &Record<'_>) -> Result<()>;
    fn flush(&self) -> Result<()>;

    // Provided methods
    fn max_log_level(&self) -> LevelFilter { ... }
    fn format(&mut self, format: FormatFunction) { ... }
    fn shutdown(&self) { ... }
    fn reopen_output(&self) -> Result<(), FlexiLoggerError> { ... }
    fn rotate(&self) -> Result<(), FlexiLoggerError> { ... }
}

Writes to a single log output stream.

Boxed instances of LogWriter can be used as additional log targets (see writers for more details).

Required Methods

fn write(&self, now: &mut DeferredNow, record: &Record<'_>) -> Result<()>

Writes out a log line.

Errors

std::io::Error

fn flush(&self) -> Result<()>

Flushes any buffered records.

Errors

std::io::Error

Provided Methods

fn max_log_level(&self) -> LevelFilter

Provides the maximum log level that is to be written.

fn format(&mut self, format: FormatFunction)

Sets the format function.

Defaults to default_format, but can be changed with a call to Logger::format_for_writer.

The default implementation is a no-op.

fn shutdown(&self)

Cleanup open resources, if necessary.

fn reopen_output(&self) -> Result<(), FlexiLoggerError>

Re-open the current output, if meaningful.

This method is called from LoggerHandle::reopen_output for all registered additional writers.

Errors

Depend on registered writers.

fn rotate(&self) -> Result<(), FlexiLoggerError>

Rotate the current output, if meaningful.

This method is called from LoggerHandle::trigger_rotation for all registered additional writers.

Errors

Depend on registered writers.